+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Changelog
 
 - Made items without use script calls always show a chat card on use (#891)
+- Added `token` to the variables of Script Calls
 
 ### Compendium